+++ src/components/popover/overlayManager.service.ts
@@ -15,7 +15,9 @@ export class OverlayManagerService {
 
     let zIndex = 100;
 
-    // this.components.filter(component => component.zIndex >= zIndex).forEach(component => zIndex=component.zIndex);
+    // this.components
+    // .filter(component => component.zIndex >= zIndex)
+    // .forEach(component => zIndex=component.zIndex);
     for (let i = 0; i < this.components.length; i++) {
       if (this.components[i].zIndex >= zIndex) {
         zIndex = this.components[i].zIndex;